requirements.txt
iiif-prezi3
shapely
scour
blend-modes
pagefind[bin]
rdflib
beautifulsoup4
pyyaml
